+++ src/test/java/com/openggf/game/rewind/TestScalarOnlyCodecDeletion.java
@@ -1227,6 +1227,24 @@ public class TestScalarOnlyCodecDeletion {
                             "com.openggf.game.sonic3k.objects.MgzMinibossInstance$MgzBossCameraScrollHelper",
                             "targetX"));
 
+    private static final List<MutableFieldCoverageCandidate> BATCH178_MUTABLE_FIELDS =
+            List.of(
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic3k.objects.CutsceneKnucklesLbz1CollapseChild",
+                            "subtype"),
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic3k.objects.CutsceneKnucklesLbz1RangeHelper",
+                            "x", "y"),
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic3k.objects.Sonic3kStarPostBonusStarChild",
+                            "centerX", "centerY"),
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic3k.objects.Sonic3kStarPostStarChild",
+                            "centerX", "centerY"),
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic3k.objects.bosses.MhzEndBossWeatherVisualChild",
+                            "spark", "subtype"));
+
     private static final List<CodecDeletionCandidate> BATCH31_DELETED_CODECS = List.of(
             new CodecDeletionCandidate(BombPrizeObjectInstance.class.getName(), GameId.S2));
 
@@ -5155,6 +5173,23 @@ public class TestScalarOnlyCodecDeletion {
     @Test
     void batch177GumballAndMgzHelperScalarsAreMutableForCompactRestore() {
         for (MutableFieldCoverageCandidate candidate : BATCH177_MUTABLE_FIELDS) {
+            Class<?> cls = loadClass(candidate.fqn());
+            for (String fieldName : candidate.fieldNames()) {
+                try {
+                    var field = findField(cls, fieldName);
+                    assertFalse(Modifier.isFinal(field.getModifiers()),
+                            cls.getName() + "#" + fieldName
+                                    + " must be mutable so compact restore can replay captured scalars");
+                } catch (NoSuchFieldException e) {
+                    throw new AssertionError("Missing scalar field " + cls.getName() + "#" + fieldName, e);
+                }
+            }
+        }
+    }
+
+    @Test
+    void batch178GraphHelperScalarsAreMutableForCompactRestore() {
+        for (MutableFieldCoverageCandidate candidate : BATCH178_MUTABLE_FIELDS) {
             Class<?> cls = loadClass(candidate.fqn());
             for (String fieldName : candidate.fieldNames()) {
                 try {